Refactor Command line game logic into separate class
Break out functions specific to the command-line operation of the game into an new class.  Keep the base Game class for future use.
