+++ connect_four.js
@@ -1,0 +1,2 @@
+const { CLIGame } = require('./lib/CLIGame');
+const game = new CLIGame().start();
