--- index.js
@@ -1,3 +1,0 @@
-const Game = require('./lib/Game');
-
-const game = new Game().start();
